Information for the results of the presidental election in Turkey by province in 2018

A dataset showing the information for the candidates competing in the presidental election, and their votes in count and percentage.

cb_2018_il

Format

A data frame with 81 rows and 105 variables:

Il_Adi

il adi, the name of the province

Kayitli_Secmen_Sayisi

kayitli secmen sayisi, the number of voters

Oy_Kullanan_Secmen_Sayisi

oy kullanan secmen sayisi, the number of voters who vote

Gecerli_Oy_Toplami

gecerli oy sayisi, the number of valid votes

MUHARREM_INCE

muharrem incenin aldigi oy sayisi, the number of valid votes that muharrem ince received

MERAL_AKSENER

meral aksenerin aldigi oy sayisi, the number of valid votes that meral aksener received

RECEP_TAYYIP_ERDOGAN

recep tayyip erdoganin aldigi oy sayisi, the number of valid votes that recep tayyip erdogan received

SELAHATTIN_DEMIRTAS

selahattin demirtasin aldigi oy sayisi, the number of valid votes that selahattin demirtas received

TEMEL_KARAMOLLAOGLU

temel karamollaoglunun aldigi oy sayisi, the number of valid votes that temel karamollaoglu received

DOGU_PERINCEK

dogu perincekin aldigi oy sayisi, the number of valid votes that dogu perincek received

MUHARREM_INCE_%

muharrem incenin aldigi oy yuzdesi, the percentage of valid votes that muharrem ince received

MERAL_AKSENER_%

meral aksenerin aldigi oy yuzdesi, the percentage of valid votes that meral aksener received

RECEP_TAYYIP_ERDOGAN_%

recep tayyip erdoganin aldigi oy yuzdesi, the percentage of valid votes that recep tayyip erdogan received

SELAHATTIN_DEMIRTAS_%

selahattin demirtasin aldigi oy yuzdesi, the percentage of valid votes that selahattin demirtas received

TEMEL_KARAMOLLAOGLU_%

temel karamollaoglunun aldigi oy yuzdesi, the percentage of valid votes that temel karamollaoglu received

DOGU_PERINCEK_%

dogu perincekin aldigi oy yuzdesi, the percentage of valid votes that dogu perincek received

Source

YSK (Supreme Election Council)