data_list

list the name of the data sets with the election names

data_list()

Value

A data frame representing the file names and election names

Examples

# \donttest{
data_list()
#>                                                   election_names
#> 1            2014 Cumhurbaskanligi Secimi (Illere gore sonuclar)
#> 2                  2014 Cumhurbaskanligi Secimi (Turkiye geneli)
#> 3            2018 Cumhurbaskanligi Secimi (Illere gore sonuclar)
#> 4                  2018 Cumhurbaskanligi Secimi (Turkiye geneli)
#> 5      2023 Cumhurbaskanligi Secimi 2.Tur (Illere gore sonuclar)
#> 6            2023 Cumhurbaskanligi Secimi 2.Tur (Turkiye geneli)
#> 7      2023 Cumhurbaskanligi Secimi 1.Tur (Illere gore sonuclar)
#> 8            2023 Cumhurbaskanligi Secimi 1.Tur (Turkiye geneli)
#> 9          1995 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 10               1995 Milletvekili Genel Secimi (Turkiye geneli)
#> 11         1999 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 12               1999 Milletvekili Genel Secimi (Turkiye geneli)
#> 13         2002 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 14               2002 Milletvekili Genel Secimi (Turkiye geneli)
#> 15         2007 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 16               2007 Milletvekili Genel Secimi (Turkiye geneli)
#> 17         2011 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 18               2011 Milletvekili Genel Secimi (Turkiye geneli)
#> 19         2018 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 20               2018 Milletvekili Genel Secimi (Turkiye geneli)
#> 21         2023 Milletvekili Genel Secimi (Illere gore sonuclar)
#> 22               2023 Milletvekili Genel Secimi (Turkiye geneli)
#> 23 2015 Haziran Milletvekili Genel Secimi (Illere gore sonuclar)
#> 24       2015 Haziran Milletvekili Genel Secimi (Turkiye geneli)
#> 25   2015 Kasim Milletvekili Genel Secimi (Illere gore sonuclar)
#> 26         2015 Kasim Milletvekili Genel Secimi (Turkiye geneli)
#> 27              2004 Belediye Baskanligi Secimi (Turkiye geneli)
#> 28              2009 Belediye Baskanligi Secimi (Turkiye geneli)
#> 29   2009 Buyuksehir Belediye Baskanligi Secimi (Turkiye geneli)
#> 30        2009 Belediye Baskanligi Secimi (Illere gore sonuclar)
#> 31        2014 Belediye Baskanligi Secimi (Illere gore sonuclar)
#> 32              2014 Belediye Baskanligi Secimi (Turkiye geneli)
#> 33        2019 Belediye Baskanligi Secimi (Illere gore sonuclar)
#> 34              2019 Belediye Baskanligi Secimi (Turkiye geneli)
#>                              data_name
#> 1                           cb_2014_il
#> 2                      cb_2014_turkiye
#> 3                           cb_2018_il
#> 4                      cb_2018_turkiye
#> 5                cb_2023_ikinci_tur_il
#> 6           cb_2023_ikinci_tur_turkiye
#> 7                   cb_2023_ilk_tur_il
#> 8              cb_2023_ilk_tur_turkiye
#> 9                  genel_secim_1995_il
#> 10            genel_secim_1995_turkiye
#> 11                 genel_secim_1999_il
#> 12            genel_secim_1999_turkiye
#> 13                 genel_secim_2002_il
#> 14            genel_secim_2002_turkiye
#> 15                 genel_secim_2007_il
#> 16            genel_secim_2007_turkiye
#> 17                 genel_secim_2011_il
#> 18            genel_secim_2011_turkiye
#> 19                 genel_secim_2018_il
#> 20            genel_secim_2018_turkiye
#> 21                 genel_secim_2023_il
#> 22            genel_secim_2023_turkiye
#> 23         genel_secim_haziran_2015_il
#> 24    genel_secim_haziran_2015_turkiye
#> 25           genel_secim_kasim_2015_il
#> 26      genel_secim_kasim_2015_turkiye
#> 27 yerel_secim_2004_buyuksehir_turkiye
#> 28   yerel_secim_2009_belediye_turkiye
#> 29 yerel_secim_2009_buyuksehir_turkiye
#> 30                 yerel_secim_2009_il
#> 31 yerel_secim_2014_buyuksehir_turkiye
#> 32                 yerel_secim_2014_il
#> 33 yerel_secim_2019_buyuksehir_turkiye
#> 34                 yerel_secim_2019_il
# }