Information for the results of the general election in Turkey by province in 2023

A dataset showing the information for each political party and alliance competing in the general election, their individual vote counts, percentages, and the overall voter statistics by province.

genel_secim_2023_il

Format

A data frame with 81 rows and 63 variables:

Il

il, the name of the province

Kayitli_Secmen_Sayisi

kayitli secmen sayisi, the number of registered voters

Oy_Kullanan_Secmen_Sayisi

oy kullanan secmen sayisi, the number of voters who vote

Gecerli_Oy_Toplami

gecerli oy toplami, the total number of valid votes

MILLET

millet partisinin oy sayisi, the number of votes the Nation Party received

HAK-PAR

hak-par'in oy sayisi, the number of votes the Rights and Freedoms Party received

TKP

tkp'nin oy sayisi, the number of votes the Communist Party of Turkey received

TKH

tkh'nin oy sayisi, the number of votes the People's Liberation Party received

SOL_PARTI

sol parti'nin oy sayisi, the number of votes the Left Party received

GENCPARTI

gencparti'nin oy sayisi, the number of votes the Young Party received

MEMLEKET

memleket partisinin oy sayisi, the number of votes the Homeland Party received

BBP

bbp'nin oy sayisi, the number of votes the Great Unity Party received

AK_Parti

ak parti'nin oy sayisi, the number of votes the Justice and Development Party received

YENIDEN_REFAH

yeniden refah partisinin oy sayisi, the number of votes the Refah Party received

MHP

mhp'nin oy sayisi, the number of votes the Nationalist Movement Party received

YESIL_SOL_PARTI

yesil sol partinin oy sayisi, the number of votes the Green Left Party received

TIP

tip'in oy sayisi, the number of votes the Labour Party received

AB

ab'nin oy sayisi, the number of votes the Bloc Party received

ANAP

anap'in oy sayisi, the number of votes the Motherland Party received

YP

yp'nin oy sayisi, the number of votes the New Path Party received

HKP

hkp'nin oy sayisi, the number of votes the People's Liberation Party received

MILLI_YOL

milli yol'un oy sayisi, the number of votes the National Course Party received

VATAN_PARTISI

vatan partisinin oy sayisi, the number of votes the Homeland Party received

GBP

gbp'nin oy sayisi, the number of votes the Independent Turkey Party received

CHP

chp'nin oy sayisi, the number of votes the Republican People's Party received

IYI_Parti

iyi parti'nin oy sayisi, the number of votes the Good Party received

AP

ap'nin oy sayisi, the number of votes the Alternative Party received

ZAFER_PARTISI

zafer partisinin oy sayisi, the number of votes the Victory Party received

SOSYALIST_GUC_BIRLIGI_ITTIFAKI

sosyalist guc birligi ittifakinin oy sayisi, the number of votes the Socialist Power Union Alliance received

CUMHUR_ITTIFAKI

cumhur ittifakinin oy sayisi, the number of votes the People's Alliance received

EMEK_VE_OZGURLUK_ITTIFAKI

emek ve ozgurluk ittifakinin oy sayisi, the number of votes the Labour and Freedom Alliance received

MILLET_ITTIFAKI

millet ittifakinin oy sayisi, the number of votes the Nation Alliance received

ATA_ITTIFAKI

ata ittifakinin oy sayisi, the number of votes the Ancestor Alliance received

BAGIMSIZ_TOPLAM_OY

bagimsiz adaylarin toplam oy sayisi, the total number of votes independent candidates received

MILLET_%

millet partisinin oy yuzdesi, the percentage of votes the Nation Party received

HAK-PAR_%

hak-par'in oy yuzdesi, the percentage of votes the Rights and Freedoms Party received

TKP_%

tkp'nin oy yuzdesi, the percentage of votes the Communist Party of Turkey received

TKH_%

tkh'nin oy yuzdesi, the percentage of votes the People's Liberation Party received

SOL_PARTI_%

sol parti'nin oy yuzdesi, the percentage of votes the Left Party received

GENCPARTI_%

gencparti'nin oy yuzdesi, the percentage of votes the Young Party received

MEMLEKET_%

memleket partisinin oy yuzdesi, the percentage of votes the Homeland Party received

BBP_%

bbp'nin oy yuzdesi, the percentage of votes the Great Unity Party received

AK_Parti_%

ak parti'nin oy yuzdesi, the percentage of votes the Justice and Development Party received

YENIDEN_REFAH_%

yeniden refah partisinin oy yuzdesi, the percentage of votes the Refah Party received

MHP_%

mhp'nin oy yuzdesi, the percentage of votes the Nationalist Movement Party received

YESIL_SOL_PARTI_%

yesil sol partinin oy yuzdesi, the percentage of votes the Green Left Party received

TIP_%

tip'in oy yuzdesi, the percentage of votes the Labour Party received

AB_%

ab'nin oy yuzdesi, the percentage of votes the Bloc Party received

ANAP_%

anap'in oy yuzdesi, the percentage of votes the Motherland Party received

YP_%

yp'nin oy yuzdesi, the percentage of votes the New Path Party received

HKP_%

hkp'nin oy yuzdesi, the percentage of votes the People's Liberation Party received

MILLI_YOL_%

milli yol'un oy yuzdesi, the percentage of votes the National Course Party received

VATAN_PARTISI_%

vatan partisinin oy yuzdesi, the percentage of votes the Homeland Party received

GBP_%

gbp'nin oy yuzdesi, the percentage of votes the Independent Turkey Party received

CHP_%

chp'nin oy yuzdesi, the percentage of votes the Republican People's Party received

IYI_Parti_%

iyi parti'nin oy yuzdesi, the percentage of votes the Good Party received

AP_%

ap'nin oy yuzdesi, the percentage of votes the Alternative Party received

ZAFER_PARTISI_%

zafer partisinin oy yuzdesi, the percentage of votes the Victory Party received

SOSYALIST_GUC_BIRLIGI_ITTIFAKI_%

sosyalist guc birligi ittifakinin oy yuzdesi, the percentage of votes the Socialist Power Union Alliance received

CUMHUR_ITTIFAKI_%

cumhur ittifakinin oy yuzdesi, the percentage of votes the People's Alliance received

EMEK_VE_OZGURLUK_ITTIFAKI_%

emek ve ozgurluk ittifakinin oy yuzdesi, the percentage of votes the Labour and Freedom Alliance received

MILLET_ITTIFAKI_%

millet ittifakinin oy yuzdesi, the percentage of votes the Nation Alliance received

ATA_ITTIFAKI_%

ata ittifakinin oy yuzdesi, the percentage of votes the Ancestor Alliance received

BAGIMSIZ_TOPLAM_OY_%

bagimsiz adaylarin toplam oy yuzdesi, the percentage of votes independent candidates received

Source

YSK (Supreme Election Council)