Information for the results of the local elections in Turkey by province in 2014, detailing votes for each party, including percentages

A dataset showing detailed information for each political party competing in the local elections in every province, including vote counts, vote percentages, and overall voting statistics.

yerel_secim_2014_il

Format

A data frame with rows for 81 rows and 52 variables:

Il_Adi

il adi, the name of the province

Kayitli_Secmen_Sayisi

kayitli secmen sayisi, the number of registered voters in each province

Oy_Kullanan_Secmen_Sayisi

oy kullanan secmen sayisi, the number of voters who voted in each province

Gecerli_Oy_Toplami

gecerli oy toplami, the total number of valid votes in each province

HUDA_PAR

huda-par'in oy sayisi, the number of votes the Free Cause Party (HUDA-PAR) received in each province

HUDA_PAR_%

huda-par'in oy yuzdesi, the percentage of votes the Free Cause Party (HUDA-PAR) received in each province

DSP

dsp'nin oy sayisi, the number of votes the Democratic Left Party (DSP) received in each province

DSP_%

dsp'nin oy yuzdesi, the percentage of votes the Democratic Left Party (DSP) received in each province

DYP

dyp'nin oy sayisi, the number of votes the True Path Party (DYP) received in each province

DYP_%

dyp'nin oy yuzdesi, the percentage of votes the True Path Party (DYP) received in each province

TURK_PARTI

turk partisi'nin oy sayisi, the number of votes the Turkish Party received in each province

TURK_PARTI_%

turk partisi'nin oy yuzdesi, the percentage of votes the Turkish Party received in each province

HKP

hkp'nin oy sayisi, the number of votes the People's Liberation Party (HKP) received in each province

HKP_%

hkp'nin oy yuzdesi, the percentage of votes the People's Liberation Party (HKP) received in each province

TKP

tkp'nin oy sayisi, the number of votes the Communist Party of Turkey (TKP) received in each province

TKP_%

tkp'nin oy yuzdesi, the percentage of votes the Communist Party of Turkey (TKP) received in each province

BBP

bbp'nin oy sayisi, the number of votes the Great Unity Party (BBP) received in each province

BBP_%

bbp'nin oy yuzdesi, the percentage of votes the Great Unity Party (BBP) received in each province

AK_PARTI

ak parti'nin oy sayisi, the number of votes the Justice and Development Party (AK PARTI) received in each province

AK_PARTI_%

ak parti'nin oy yuzdesi, the percentage of votes the Justice and Development Party (AK PARTI) received in each province

YURT-P

yurt-p'nin oy sayisi, the number of votes the Homeland Party (YURT-P) received in each province

YURT-P_%

yurt-p'nin oy yuzdesi, the percentage of votes the Homeland Party (YURT-P) received in each province

DP

dp'nin oy sayisi, the number of votes the Democratic Party (DP) received in each province

DP_%

dp'nin oy yuzdesi, the percentage of votes the Democratic Party (DP) received in each province

EMEP

emep'in oy sayisi, the number of votes the Labour Party (EMEP) received in each province

EMEP_%

emep'in oy yuzdesi, the percentage of votes the Labour Party (EMEP) received in each province

ODP

odp'nin oy sayisi, the number of votes the Freedom and Solidarity Party (ODP) received in each province

ODP_%

odp'nin oy yuzdesi, the percentage of votes the Freedom and Solidarity Party (ODP) received in each province

MILLET

millet partisinin oy sayisi, the number of votes the Nation Party (MILLET) received in each province

MILLET_%

millet partisinin oy yuzdesi, the percentage of votes the Nation Party (MILLET) received in each province

SAADET

saadet partisinin oy sayisi, the number of votes the Felicity Party (SAADET) received in each province

SAADET_%

saadet partisinin oy yuzdesi, the percentage of votes the Felicity Party (SAADET) received in each province

HAK-PAR

hak-par'in oy sayisi, the number of votes the Rights and Freedoms Party (HAK-PAR) received in each province

HAK-PAR_%

hak-par'in oy yuzdesi, the percentage of votes the Rights and Freedoms Party (HAK-PAR) received in each province

LDP

ldp'nin oy sayisi, the number of votes the Liberal Democratic Party (LDP) received in each province

LDP_%

ldp'nin oy yuzdesi, the percentage of votes the Liberal Democratic Party (LDP) received in each province

BDP

bdp'nin oy sayisi, the number of votes the Peace and Democracy Party (BDP) received in each province

BDP_%

bdp'nin oy yuzdesi, the percentage of votes the Peace and Democracy Party (BDP) received in each province

BTP

btp'nin oy sayisi, the number of votes the Independent Turkey Party (BTP) received in each province

BTP_%

btp'nin oy yuzdesi, the percentage of votes the Independent Turkey Party (BTP) received in each province

IP

ip'nin oy sayisi, the number of votes the Workers' Party (IP) received in each province

IP_%

ip'nin oy yuzdesi, the percentage of votes the Workers' Party (IP) received in each province

HDP

hdp'nin oy sayisi, the number of votes the Peoples' Democratic Party (HDP) received in each province

HDP_%

hdp'nin oy yuzdesi, the percentage of votes the Peoples' Democratic Party (HDP) received in each province

CHP

chp'nin oy sayisi, the number of votes the Republican People's Party (CHP) received in each province

CHP_%

chp'nin oy yuzdesi, the percentage of votes the Republican People's Party (CHP) received in each province

HEPAR

hepar'in oy sayisi, the number of votes the Rights and Equality Party (HEPAR) received in each province

HEPAR_%

hepar'in oy yuzdesi, the percentage of votes the Rights and Equality Party (HEPAR) received in each province

MHP

mhp'nin oy sayisi, the number of votes the Nationalist Movement Party (MHP) received in each province

MHP_%

mhp'nin oy yuzdesi, the percentage of votes the Nationalist Movement Party (MHP) received in each province

BAGIMSIZ_TOPLAM_OY

bagimsiz adaylarin toplam oy sayisi, the total number of votes received by independent candidates in each province

BAGIMSIZ_TOPLAM_OY_%

bagimsiz adaylarin toplam oy yuzdesi, the total percentage of votes received by independent candidates in each province

Source

YSK (Supreme Election Council)